gemini-gradio-poc/requirements.txt
google-generativeai>=0.3.0
gradio>=4.0.0
pandas>=1.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
python-docx>=0.8.10
PyPDF2>=3.0.0
python-dotenv>=1.0.0